feat: title value lookups for Table and MarkupTable to allow for using choices on columns, re-used for FactSet
