+++ lib/builders/FactSet.js
@@ -1,5 +1,6 @@
 const ComponentBuilder = require('./../utils/Component-builder')
 const applyBadge = require('./../utils/apply-badge')
+const lookupOrValue = require('./../utils/lookup-or-value')
 
 const MARGINS = {
   small: 'sm',
@@ -98,18 +99,4 @@ module.exports = function (definition, options) {
   }
 
   return builder.compile()
-}
-
-function lookupOrValue (title, value, choices) {
-  if (Array.isArray(choices) === true) {
-    const data = value.replace('{{', '').replace('}}', '')
-
-    const arrVal = `${data}.map(d => lists.$simpleTitleMaps['${title}'][d] || d || '-').join(', ')`
-    const strVal = `lists.$simpleTitleMaps['${title}'][${data}] || ${data}`
-    const defaultVal = '-'
-
-    return `{{ ( Array.isArray(${data}) ? ${arrVal} : (${strVal}) ) || '${defaultVal}' }}`
-  } else {
-    return value
-  }
 }
