+++ lib/builders/MarkupTable.js
@@ -3,6 +3,7 @@ const applySpacing = require('./../utils/apply-spacing')
 const applyTooltip = require('./../utils/apply-tooltip')
 const applyBadge = require('./../utils/apply-badge')
 const applyLaunches = require('./../utils/apply-launches')
+const lookupOrValue = require('./../utils/lookup-or-value')
 const { COLORS, TEXT_WHITE } = require('./../utils/color-reference')
 
 module.exports = function (definition, options) {
@@ -131,11 +132,13 @@ function horizontal (qMarkupTable, definition) {
     applySelection(selectionTd, definition, destPath)
   }
 
-  for (const { value, showWhen, badge, tooltip, showLaunches } of columns) {
+  for (const { title, value, choices, showWhen, badge, tooltip, showLaunches } of columns) {
+    const display = lookupOrValue(title, value, choices)
+
     const cell = tBodyRow
       .addChildTag('td')
       .addAttribute('class', 'text-left')
-      .content(value)
+      .content(display)
 
     if (showWhen) cell.addAttribute('v-if', showWhen)
     if (tooltip) applyTooltip({ source: cell, tooltip, icon: false })
@@ -218,7 +221,7 @@ function vertical (qMarkupTable, definition) {
 
   let colIdx = 0
 
-  for (const { title, value, showWhen, badge, tooltip, showLaunches, titleTooltip, color } of columns) {
+  for (const { title, value, choices, showWhen, badge, tooltip, showLaunches, titleTooltip, color } of columns) {
     const tr = tBody.addChildTag('tr')
 
     const trClasses = []
@@ -242,12 +245,14 @@ function vertical (qMarkupTable, definition) {
       applyVerticalHeader(tr, { title, titleTooltip, fixedHeader, colIdx })
     }
 
+    const display = lookupOrValue(title, value, choices)
+
     const cell = tr
       .addChildTag(fixedHeader && colIdx === 0 ? 'th' : 'td')
       .addAttribute('class', 'text-left')
       .addAttribute('v-for', `(item, rowIdx) in ${arrayPath}`)
       .addAttribute(':key', 'rowIdx')
-      .content(value)
+      .content(display)
 
     const cellStyles = []
