+++ lib/builders/Table.js
@@ -1,8 +1,9 @@
 const ComponentBuilder = require('./../utils/Component-builder')
 const applySpacing = require('./../utils/apply-spacing')
+const applyLaunches = require('./../utils/apply-launches')
+const lookupOrValue = require('./../utils/lookup-or-value')
 const { COLORS, TEXT_WHITE } = require('./../utils/color-reference')
 const safeInput = str => str.replace(/'/g, '\\\'')
-const applyLaunches = require('./../utils/apply-launches')
 
 module.exports = function (definition, options) {
   const {
@@ -109,9 +110,15 @@ module.exports = function (definition, options) {
       .addChildTag('q-td')
       .addAttribute('key', col.field)
 
+    const value = `{{ props.cols[${idx}].value }}`
+
+    const display = col.choices
+      ? lookupOrValue(col.title, value, col.choices)
+      : value
+
     td
       .addChildTag('span')
-      .content(`{{ props.cols[${idx}].value }}`)
+      .content(display)
 
     if (col.flag) {
       const chip = td
@@ -154,5 +161,6 @@ module.exports = function (definition, options) {
 
   applySpacing({ spacing, source: table })
 
+  console.log(builder.compile())
   return builder.compile()
 }
